Fixed transparency issue.
The artifacting turned out to be an issue with an incorrectly set DepthStencilState.
During the bug hunt, I noticed that the Camera LookAt matrix was being constructed incorrectly.
Added a renderer to show the vertex normals when the Mesh is drawn in Wireframe mode.
Corrected the height map matrix indexing.
Added a "walk mode" for movement. Holding shift while moving slows things to 10%.