Commit of work so far done to convert the project into this century and attempt to hopefully clean out alot of horrible old code.
E.g the sqlutil etc and any other horrible code, and eventually end up cleaning away the various horrible attributes etc and update column names e.g AccountId to Id etc
Need to sort out the files still in the sense that they are still full of various methods and other gubbins which shouldn't be in the entity file at all.
Overall project cleanup and code update I think is required as much of it seems insanely over complicated but also messy and dangerous compared to what it needs to/should be.
Maybe out of this we can also put a cap on the mad resource usage associated with WCell as well.

NOTE: Databaseprovider as it is now is highly likely to be a memory leak and needs correcting, however it is in a "it works" state.